+++ shared/@common/components/ui/modal/components/index.ts
@@ -5,7 +5,7 @@ import OutlineButton from "./modal-compound-outline-button"
 import PrimaryButton from "./modal-compound-primary-button"
 import DeleteButton from "./modal-compound-delete-button"
 import Form from "./modal-compound-form"
-import Layout from "./modal-compound-layout"
+import Contents from "./modal-compound-contents"
 import Description from "./modal-compound-description"
 import Close from "./modal-compound-close"
 import Header from "./modal-compound-header"
@@ -18,7 +18,7 @@ export {
   PrimaryButton,
   DeleteButton,
   Form,
-  Layout,
+  Contents,
   Description,
   Close,
   Header,
