+++ shared/@common/components/ui/modal/components/modal-compound-button-layout.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from "react"
-import classes from "./modal.module.css"
+import classes from "../index.module.css"
 
 export default function ButtonLayout(props: PropsWithChildren) {
   return <div className={classes["button-layout"]}>{props.children}</div>
