+++ shared/@common/components/ui/modal/components/modal-compound-description.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from "react"
-import classes from "./modal.module.css"
+import classes from "../index.module.css"
 
 export default function Description(props: PropsWithChildren) {
   return <p className={classes.description}>{props.children}</p>
