+++ shared/@common/components/ui/modal/components/modal-compound-title.tsx
@@ -1,5 +1,5 @@
 import { useModalContext } from ".."
-import classes from "./modal.module.css"
+import classes from "../index.module.css"
 
 export default function Title() {
   const modalContext = useModalContext()
