+++ shared/@common/components/ui/modal/index.module.css
@@ -1,4 +1,4 @@
-.modal-layout {
+.modal-root {
   padding: 32px 28px;
   background-color: white;
   position: fixed;
@@ -9,6 +9,10 @@
   border-radius: 8px;
   max-height: 95vh;
   overflow-y: scroll;
+}
+
+.modal-contents {
+  position: relative;
 }
 
 .small {
